requirements.txt
Pillow>=2.3.0
numpy>=1.7
scipy>=0.13.3
protobuf>=2.5.0
Pillow>=2.3.0
numpy>=1.8.2
scipy>=0.13.3
protobuf>=2.5.0
six>=1.5.2
requests>=2.2.1
gevent>=1.0
Flask>=0.10.1
Flask-WTF>=0.11
Flask-SocketIO==0.6.0
gunicorn==17.5
setuptools>=3.3
lmdb>=0.87
h5py>=2.2.1
pydot2
